Good News Everybody!
I've developed the connection classes to the point where you can call getBroker() and issue an SqlStatement, have it sent into a queue which is read by a thread that is then sent to the database. We are on track!
I still need to have actual statements to test on the database, as well as load testing, but now the structure is solidifying and we will soon have actual tangible applications for our work. Give me another day or two and we should have a solid db connection. Actual Broker cache work is lagging a little, but once connectivity is good to go it should be a few hours per broker. If we are behind on schedule, it shouldn't be by too much. 

Good work all!