+ Reworked the database connections to follow the simplified setup I mentioned in this morning's class.  Let me know what you think! This is just the basic, stripped-down framework.
For this to work, you MUST first rerun the DB Create script (Filled version) which now includes a new user that brokers will use for DB access.

You can check out the TestEmpSearch class in the test folder for an example of how the broker would be used when searching for employees. It runs as-is. Make sure your MySQL instance/service is running. :)

Thanks to SVN we can still revert to the old setup whenever needed, just by rolling back to the revision before this one.

Some additional notes:
- All previous db classes moved to the oldClasses* packages.  There's still a lot of good usable stuff in them. :D

- As mentioned, the DB creation scripts have been updated to include a new broker user, which for now only has access to the employee table. Access will be expanded to other tables as needed.

- I wanted to get this commit out right away, so currently the brokers lack the ability to close old/inactive db connections.  That will be added soon, in addition to add/update/delete support for employees.