Updates and Deletes work for permission levels!
OK, I think we need to make a major update to our permissionbroker:
We need to move it to the  permissions package. It's more secure, it's proper, and we can
always have a PermissionBroker in persistance that basically returns the other PermissionBroker object.
I think this is the best, most secure way to deal with permissions. It might require some structural work,
but I'm willing to do that.

Will talk about it in class. Cheers.