Big update.
Broker structure changed to account for "race conditions". Updates now need the old/original object to be passed in, in addition to the updated object. Deletes now require an object from the database as well, not just an object with the employee ID to delete.

Database edited to remove need for the EmployeeRecordID. Now based off of the Employee ID again.  EmployeeBroker modified accordingly as JOINs in statements are no longer needed.

LocationBroker's update and delete methods have already been updated to match the new broker structure. TestLocationBroker class is likely to be broken, but will be fixed in a few minutes to run the tests properly.