Added Backup class and method, Backup.backupDB(), which will create full backups of the WebAgenda database.
You will need to run the create script again for this to work, as the WABroker user needed new privilidges.

For you to test this, you'll have to change the main() method inside the class to point to where your MySQL bin directory is, and where you want to output backups to.